Split platform-specific parts of RenderTargetCube into separate files
