+++ MonoGame.Framework/Graphics/RenderTargetCube.OpenGL.cs
@@ -1,0 +1,16 @@
+﻿// MonoGame - Copyright (C) The MonoGame Team
+// This file is subject to the terms and conditions defined in
+// file 'LICENSE.txt', which is part of this source code package.
+
+using System;
+
+namespace Microsoft.Xna.Framework.Graphics
+{
+    public partial class RenderTargetCube
+    {
+        private void PlatformConstruct(GraphicsDevice graphicsDevice, SurfaceFormat preferredFormat, DepthFormat preferredDepthFormat, int preferredMultiSampleCount)
+        {
+            throw new NotImplementedException();
+        }
+    }
+}
